[*] now system generates genesis block and persists it to blockchain with executing transactions. [=] fixed error with storing global blockchain configuration (created new protobuf structure for it). [*] many different bug fixes and code improvements.

+++ Phorkus/Phorkus.Core/Storage/IBlockRepository.cs
@@ -10,7 +10,7 @@ namespace Phorkus.Core.Storage
         /// </summary>
         /// <param name="blockHeight">The block height / index to retrieve</param>
         /// <returns>Block hash at specified height / index</returns>
-        Block GetBlockByHeight(uint blockHeight);
+        Block GetBlockByHeight(ulong blockHeight);
         
         /// <summary>
         /// Retrieves a block by hash
